gnf to dds support

+++ inc/converter.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "pch.h"
 
-int ConvertGnfToDDS(const byte* gnfsrc, byte*& ddsout);
+size_t ConvertGnfToDDS(const byte* gnfsrc, byte*& ddsout);
